Catch exceptions and deal with them and append instead of write files
So now if an output file doesn't exist, it'll be created, but if it
does, it'll just be appended to instead of being overwritten.

For exceptions, now whenever there is an exception (the UTF-8 encoding
error bullshit), the exception block will delete all lines (in the input
file) up to the error line, and then just resume parsing. So the input
file will be all fucked up by the end of running the converter, but the
output should be produced relatively quickly and without any hiccoughs.
Maybe not ideal, but it works for now.