Comment out columns that are not needed; switch position of two columns
I just commented out the entries instead of removing them because
creating these arrays was painstaking effort and was created by hand by
looking at the spec online. I would not want to re type this shit if I
discover that I might need a new column in the future. Also, the order
matters, so it's best that they all stay in there.

I also switched the positions of Street Number and Street Name because
for some reason in the original spec the street number comes after the
street name which is super annoying. I might merge them at some point
too.