Update README to reflect automatic deletion of foul lines and clean up
The README now mentions that the parser can automatically delete lines
that have a UTF-8 encoding error (CSV::MalformedCSVError), but it still
has the original instructions to do this by hand if necessary.

THe README also states that the program has to be run from the data
folder.

A stale checkbox item (better known as a task) has also been removed.