Changes
This is a horrible commit, but I want to have some sort of backup for
the breakthrough I just made. This is pretty much done at this point. I
just need to clean it up and add what I need to and stuff.

By pretty much done, I mean the core logic and figuring out of how the
shit is going to work. Everything else is just mundane at this point.