Add .ruby-version
I have tested this application with Ruby 2.6.2 I believe, and I'm
putting this in right now and I haven't tested anything after adding
this file and making this change. If stuff doesn't work, go back to
2.6.2 by changing this file and see if that helps.

Also, prepend commands with `bundle exec` so that the correct ruby
version is used.