Add error handling
Sometimes the Google Maps API doesn't come back with a Zip Code. Before
the entire ugly Hash that came back as a response was being appended to
the sheet. Now it'll just be an empty string.

I also wasn't handling error messages in the response back from the
Google API. That's handled now.