Put a space instead of an empty string if no Google API Data Returned
This way I can rely on Lead Sherpa's Geocoding (they geocode zip codes
if there's a filler in the cell) if the Google API is not working for
whatever reason. As of today, I haven't put in billing information and
the trial expired and so I can't use Google's API for geocoding.